refactor(ticket-management): moved ticket management into profile management
