+++ src/app/dashboard/profile-management/_components/profile-tracker-management-board.tsx
@@ -21,8 +21,8 @@ import { api } from "@/trpc/react";
 import { ChevronDownIcon } from "@radix-ui/react-icons";
 import { Loader2 } from "lucide-react";
 import UserDetailView from "./user-detail-view";
-import CustomerSelect from "../../ticket-management/_components/customer-select";
-import { CustomerList } from "../../ticket-management/_components/ticket-management-board";
+import CustomerSelect from "./customer-select";
+import { CustomerList } from "./customer-list";
 
 // flag to move back to select customer
 const DYNAMIC_CUSTOMER_ROUTE_FLAG = true;
